Syntax problem solved

+++ SQL_command/CreateTable.sql
@@ -1,5 +1,8 @@
 CREATE TABLE IF NOT EXISTS Product(
-product_id INTEGER(100) NOT NULL AUTO_INCREMENT,
-product_name VARCHAR(30),
-product_price INTEGER(300),
-PRIMARY KEY (product_id);
+    product_id INTEGER(100) NOT NULL AUTO_INCREMENT,
+    product_name VARCHAR(30),
+    product_price INTEGER(255),
+    PRIMARY KEY (product_id)
+    );
+
+
